+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   More sensible default build path
+-   Show notes in Build Actions list view
 
 ## [3.0.0] - 2021-10-19